Serve plugin manifest from Limitless plugin
